DEP-62 fix: add feignClient for oauth
