--- app/src/main/java/com/depromeet/threedays/front/client/GoogleOAuthClient.java
@@ -1,19 +1,0 @@
-package com.depromeet.threedays.front.client;
-
-import com.depromeet.threedays.front.client.model.OAuthInfo;
-import com.depromeet.threedays.front.client.property.OAuthProperty;
-import com.depromeet.threedays.front.domain.model.Token;
-import org.springframework.stereotype.Component;
-
-@Component
-public class GoogleOAuthClient implements OAuthClient {
-
-	public GoogleOAuthClient() {
-		super();
-	}
-
-	@Override
-	public OAuthInfo getOAuthInfo(OAuthProperty oAuthProperty, Token token) {
-		return null;
-	}
-}
