+++ app/src/main/java/com/depromeet/threedays/front/config/FeignConfig.java
@@ -1,0 +1,8 @@
+package com.depromeet.threedays.front.config;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@EnableFeignClients(basePackages = "com.depromeet.threedays.front.client")
+@Configuration
+public class FeignConfig {}
