+++ app/src/main/resources/messages/member.properties
@@ -1,1 +1,2 @@
-member.constraints.invalid.member = "?? ??? ???? ????."
+member.constraints.invalid.member = "\uC720\uD6A8\uD558\uC9C0 \uC54A\uC740 \uBA64\uBC84\uC785\uB2C8\uB2E4."
+4001 = "\uC694\uCCAD \uC8FC\uC18C\uAC00 \uC720\uD6A8\uD558\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4."
